lib/fasttime: remove performance overhead when UnixTimestamp() is called from benchmark loops
Move the synctest-related implementation into a separate file protected with go:build goexperiment.synctest.

This is a follow-up for the commit 06c26315df5bce973865d931c1e8489467a7e770
Updates https://github.com/VictoriaMetrics/VictoriaMetrics/pull/8740
